rest/requirements.txt
django
coverage
psycopg2-binary
gunicorn
tblib
djangorestframework
django-cors-headers
django-filter
tqdm
djangorestframework-recursive
torch
pillow
torchvision
annoy
requests
scipy
drf-yasg